pull後のconflictが修正できないので新ブランチ作成
