+++ site/templates/home.html
@@ -14,7 +14,7 @@
 </head>
 
 <body>
-    <div id="nav">
+    <div class="nav">
         <div class="container">
             <div class="container2">
                 <a href="{% url 'home' %}">芝浦課題管理システム</a>
@@ -79,9 +79,9 @@
             </div>
         </div>
     </main>
-
-    <script type="text/javascript" src="{%static 'css/js/caleandar.js'%}"></script>
-    <script type="text/javascript" src="{%static 'css/js/demo.js' %}"></script>
+    <br><br><br>
+    <script type="text/javascript" src="{%static 'js/caleandar.js'%}"></script>
+    <script type="text/javascript" src="{%static 'js/demo.js' %}"></script>
 </body>
 
 </html>
