+++ site/templates/index.html
@@ -11,7 +11,7 @@
     <link rel="stylesheet" href="{% static 'css/home.css' %}">
 </head>
 <body>
-    <div id="nav">
+    <div class="nav">
         <div class="container">
             <div class="container2">
                 <a href="http://127.0.0.1:8000/">芝浦課題管理システム</a>
