+++ site/templates/kadai.html
@@ -11,7 +11,7 @@
 	<link rel="stylesheet" href="{% static 'css/home.css' %}">
 </head>
 <body>
-    <div id="nav">
+    <div class="nav">
         <div class="container">
             <div class="container2">
                 <a href="{% url 'home' %}">芝浦課題管理システム</a>
